backend/pyproject.toml
[tool.poetry]
name = "shams_messenger"
version = "0.1.0"
description = "Messaging app"
authors = ["Adam Rashi <adamrashi@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.0"}
psycopg = {extras = ["binary"], version = "^3.1.9"}
python-dotenv = "^1.0.0"
rsa = "^4.9"
gunicorn = "^21.2.0"
alembic = "^1.11.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
